Conflicts:
	Client/Source/ClientItemHandler.cpp
	Server/GUI/Server GUI.csproj

+++ Client/Source/Game.cpp
@@ -45,8 +45,8 @@ void Game::InitGameClient(const std::string &IP, const unsigned int &port, std::
 	//Connects to a Host With the IP Address and port in the parameters
 	// TODO: RAII
 	this->zClient = new Client();
-	iImage* blackImage = GetGraphics()->CreateImage(Vector2(0,0), Vector2(GetGraphics()->GetEngineParameters().WindowWidth, 
-		GetGraphics()->GetEngineParameters().WindowHeight), "Media/LoadingScreen/FadeTexture.png");
+	iImage* blackImage = GetGraphics()->CreateImage(Vector2(0,0), Vector2( (float)GetGraphics()->GetEngineParameters().WindowWidth, 
+		(float)GetGraphics()->GetEngineParameters().WindowHeight), "Media/LoadingScreen/FadeTexture.png" );
 	this->zClient->SetBlackImage(blackImage);
 	blackImage = NULL;
 	this->zClient->Connect(IP, port, errMsg, errorCode);
